Runtime/fuzz: Rewriting the core fuzzed input generator.
The new implementation stores only changes between mutations and not the
complete input. This saves memory (at the cost of a bit of increased complexity
when generating the input) and allows to generate mutations when no initial corpus
data is available which is useful when trying to fuzz data streams like network connections.

There are two modes when creating a new fuzzing context:
     * BLOB is used for input data like files where an initial corpus is available
     * STREAM is used for data streams like network connections



git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@77482 cfe28804-0f27-0410-a406-dd0f0b0b656f