Additions: Linux: Prevent kernel modules rebuild on system boot if they already been built, bugref:9138.
This is the same approach as we have for host drivers. Before attempting to build
modules, we try to detect if modinfo can already see them and check versions. Modules
rebuilding skipped in positive case.

This change only affects 'rcvboxadd start' action which is used on system boot. Having this
change will speed up boot process a bit. Actions 'rcvboxadd setup' and 'rcvboxadd quicksetup'
not affected.


git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@94308 cfe28804-0f27-0410-a406-dd0f0b0b656f