VBoxDef2LazyLoad: Deal with DATA entries (was ignored).
The `DATA` entries cannot easily be done in a lazy nor in a transparent fashion,
since the compiler does not call any code when accessing data.  The compiler
does not necessarily generate an indirect reference to it (e.g. ELF executable
images seems to define the variable locally in the .exe image and rely on
unix-style global symbol resoltion for the .so to pick it up).  So, instead we
require special header file writing for data symbols.

Given the variable `uint64_t g_uData;` in an .so. We create the def-file line
`g_uData DATA` for it.  In the header file, we replace the
`extern uint64_t g_uData;` with `DECLASM(uint64_t *) LazyGetPtr_g_uData(void);`
and `#define g_uData (*LazyGetPtr_g_uData())`.

For code that uses the explicit load function, it would also be possible to
us the pointer variable `extern uint64_t *g_pg_uData;` directly via wrapper
`#define g_uData (*g_pg_uData)`, but do that with great care as the any
init ordering mishaps will cause accesses to `g_uData` to crash.


git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@99551 cfe28804-0f27-0410-a406-dd0f0b0b656f